requirements.txt
# Core ML libraries
pandas>=1.5.0
numpy>=1.21.0
scikit-learn>=1.1.0
xgboost>=1.6.0
lightgbm>=3.3.0
catboost>=1.1.0

# Deep Learning
tensorflow>=2.10.0
torch>=1.12.0

# Visualization and Dashboard
streamlit>=1.25.0
plotly>=5.10.0
matplotlib>=3.5.0
seaborn>=0.11.0

# API and Web
flask>=2.2.0
flask-cors>=3.0.10
fastapi>=0.85.0
uvicorn>=0.18.0

# Data and External APIs
requests>=2.28.0
python-dotenv>=0.20.0
pyyaml>=6.0

# Optimization and AutoML
optuna>=3.0.0
hyperopt>=0.2.7

# Time Series
prophet>=1.1.0
statsmodels>=0.13.0

# Utilities
tqdm>=4.64.0
joblib>=1.2.0
python-dateutil>=2.8.0

# Testing
pytest>=7.1.0
pytest-cov>=4.0.0

# Logging and Monitoring
loguru>=0.6.0
wandb>=0.13.0

# Production
gunicorn>=20.1.0
docker>=6.0.0